% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortResults.R
\docType{package}
\name{CohortResults-package}
\alias{CohortResults}
\alias{CohortResults-package}
\title{CohortResults: An R package to get summary results for one or more cohorts.}
\description{
An R package to get summary results for one or more cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortResults/}
  \item \url{https://github.com/OHDSI/CohortResults}
  \item Report bugs at \url{https://github.com/OHDSI/CohortResults/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
