% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortInclusionRules.R
\name{getCohortInclusionRules}
\alias{getCohortInclusionRules}
\title{Retrieve cohort inclusion rule statistics and results.}
\usage{
getCohortInclusionRules(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames,
  cohortIds = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTableNames}{list of table names. See CohortGenerator::getCohortTableNames}

\item{cohortIds}{A vector of cohort IDs to filter the cohort table for inclusion.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}
}
\value{
A list of tibbles containing:
\itemize{
\item \code{cohortCount}: Cohort counts as retrieved by \code{\link{getCohortCounts}}.
\item \code{inclusion}: Inclusion rules for the cohorts.
\item \code{inclusionResult}: Results for the inclusion rules.
\item \code{inclusionStats}: Statistics related to inclusion rule performance.
\item \code{summaryStats}: Summary statistics for the cohorts.
}
}
\description{
This function retrieves various inclusion rule statistics and results for the specified cohorts from the
cohort tables in a database schema.
}
\details{
This function collects information about inclusion rules, cohort counts, and statistics
from various cohort tables. It automatically handles database connections and temporary table emulation.
}
\examples{
\dontrun{
connectionDetails <- DatabaseConnector::createConnectionDetails(dbms = "postgresql",
                           server = "myserver", user = "myuser", password = "mypassword")
cohortTableNames <- list(
  cohortTable = "cohort",
  cohortInclusionTable = "cohort_inclusion",
  cohortInclusionResultTable = "cohort_inclusion_result",
  cohortInclusionStatsTable = "cohort_inclusion_stats",
  cohortSummaryStatsTable = "cohort_summary_stats"
)
inclusionRules <- getCohortInclusionRules(connectionDetails = connectionDetails,
                                         cohortDatabaseSchema = "my_schema",
                                         cohortTableNames = cohortTableNames,
                                         cohortIds = c(1, 2, 3))
}

}
